tests/wpt/web-platform-tests/workers/interfaces/WorkerGlobalScope/close/setInterval.html

<!--
var interval1 = setInterval(function() {
  clearInterval(interval1);
  postMessage(1);
  throw new Error();
}, 10);
close();
var interval2 = setInterval(function() {
  clearInterval(interval2);
  postMessage(1);
  throw new Error();
}, 10);
/*
-->
<!doctype html>
<title>close() and setInterval</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var gotMessage = false;
  var gotError = false;
  var worker = new Worker('#');
  worker.onmessage = function(e) {
    gotMessage = true;
  };
  worker.onerror = function(e) {
    gotError = true;
  };
  setTimeout(this.step_func(function() {
    assert_false(gotMessage, 'got message');
    assert_true(gotError, 'got error');
    this.done();
  }), 500);
});
</script>
<!--
*/
//-->
